pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agents"
version = "0.1.0"
description = "AI Agent Orchestration Platform"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",

    # Database
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",  # alembic needs sync driver

    # Redis
    "redis[hiredis]>=5.2.0",

    # Auth
    "pyjwt>=2.10.0",
    "passlib[bcrypt]>=1.7.4",

    # AI Providers
    "anthropic>=0.42.0",
    "openai>=1.60.0",

    # Config
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Crypto
    "cryptography>=44.0.0",

    # HTTP client (for GitHub API, self-hosted providers)
    "httpx>=0.28.0",

    # Utilities
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",  # for TestClient
    "ruff>=0.8.0",
]

[project.scripts]
agents-cli = "agents.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agents"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
